requirements.txt
fastapi
uvicorn
panda
openai
pandasai
numpy
openpyxl